Remove ProductListFragment
