+++ app/src/main/java/pl/pola_app/internal/di/PolaComponent.java
@@ -6,7 +6,6 @@ import dagger.Component;
 import pl.pola_app.PolaApplication;
 import pl.pola_app.ui.activity.MainActivity;
 import pl.pola_app.ui.fragment.ProductDetailsFragment;
-import pl.pola_app.ui.fragment.ProductsListFragment;
 import pl.pola_app.ui.fragment.ScannerFragment;
 
 @Singleton
@@ -23,6 +22,5 @@ public interface PolaComponent {
 
     void inject(MainActivity mainActivity);
     void inject(ScannerFragment scannerFragment);
-    void inject(ProductsListFragment productsListFragment);
     void inject(ProductDetailsFragment productDetailsFragment);
 }
