+++ app/src/main/res/layout/activity_main.xml
@@ -16,15 +16,15 @@
         android:layout_alignParentStart="true"
         tools:layout="@layout/fragment_scanner" />
 
-    <fragment
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/products_list"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:name="pl.pola_app.ui.fragment.ProductsListFragment"
-        android:id="@+id/product_list_fragment"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true"
         android:layout_alignParentBottom="true"
-        android:layout_alignParentRight="true"
         android:layout_alignParentEnd="true"
-        tools:layout="@layout/fragment_products_list" />
+        android:layout_alignParentLeft="true"
+        android:layout_alignParentRight="true"
+        android:layout_alignParentStart="true"
+        android:overScrollMode="never"
+        android:scrollbars="vertical"/>
 </RelativeLayout>
